Ruby API: object IDs as resource args
Allow specifying object ID as resource argument, not as a list
of arguments to action.

Example: delete IP with ID 50 from VPS ID 101.

Before:
  api.vps.ip_address.delete(101, 50)

Now:
  api.vps(101).ip_address(50).delete

It is completely optional though, it can be written even as:
  api.vps.ip_address(101, 50).delete

or use the old form. Arguments are chained and given to the
specified action, where they're applied.