Action description: add concrete links to resources, Action#prepare
When an action is asked a description with correct URL parameters
(would return an existing object), it returns description specific
to given object. Links to referenced resources are concrete, e.g.:

    "value": {
        "url": "/v1/nodes/1",
        "method": "GET",
        "help": "/v1/nodes/1?method=GET"
    }

As HaveAPI does not now how to resolve parametrized link itself,
it must be help. URL parameters for the description of referenced
resource may be specified like:

    resource MyAPI::Node, value_params: ->{ @vps.node_id  }

Similarily, there is parameter choices_params, which is used
for resolving URL of an action that returns a list of choices
for this resource.

The lambda function has to return one parameter or an array.
It is executed in the context of action instance.
For this purpose, a new method for preloading objects has been
introduced.

    class Show < HaveAPI::Actions::Default::Show
      ...

      def prepare
        @vps = Vps.find(params[:vps_id])
      end

      def exec
        all_attrs(@vps)
      end
    end

In case of action description, only method #prepare is called, #exec
is ignored. When invoking action for real, #prepare is called before

currently primarily for the lambda functions.