Model adapters
Model adapters abstract the way action input/output is handled.
ActiveRecord and Hash adapters are built-in.

ActiveRecord adapter allows actions to return simply a model instance
or a query result. The adapter will then extract attributes
from the model and prepare a hash to be encoded in a selected output
format and sent to the client.