Add support for blocking actions
The API has to implement an interface to supply a list of actions
pending completion and to provide their state. Clients can then ask
the API whether a particular action has finished, whether it was
successful or what is its progress.

Target usage is that clients will block until the action is completed,
showing its progress while its being executed.