clients/ruby: do not fail prematurely if the token expired, try it
Generally, the client does not know when and if tokens expire. For
auto renewable tokens, the server prolongs them, but does not report
for how long. If the client still has username and password, it can
request a new token, but otherwise the client should just try it and
report a real authentication failure if the token truly expired.