servers/ruby: use correct credentials for token auth examples
The examples used hardcoded `username` and `password` fields, while
these fields depend on API configuration and can be different.
Authentication examples are now generated from the description of
action Token.Request.

We're not updating haveapi-fs example, as there the authentication is
not flexible enough and only `user` option is accepted.